+++ CHANGELOG.md
@@ -68,6 +68,7 @@ v1.8.0 [unreleased]
 - [#351](https://github.com/latera/homs/pull/351) Improve claim button styles on widget form.
 - [#354](https://github.com/latera/homs/pull/354) Wrap task short pop up list page together.
 - [#353](https://github.com/latera/homs/pull/353) Integrate new claiming components.
+- [#356](https://github.com/latera/homs/pull/356) Highlight opened task in table.
 
 v1.7.8 [2019-10-03]
 -------------------